pyproject.toml
[project]
name = "cc-nim"
version = "2.0.0"
description = "Proxy that translates Claude Code CLI (Anthropic API) to NVIDIA NIM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "uvicorn>=0.34.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "websockets>=13.0",
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.12.0",
    "aiolimiter>=1.2.1",
    "openai>=2.16.0",
    "aiohttp>=3.13.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
